maintainability: decompose tracker/github.go by responsibility
github.go was 772 lines, 28 from the 800-line hard cap (scripts/file_size_budget_test.go,
empty baseline) with 15 commits in 4 weeks — split it pre-emptively before a
feature PR is forced into a rushed mid-stream split (#821, #792 follow-up).

Move by responsibility seam (no renames, no behavior change), keeping the
client core (struct, constructors, wire types) in github.go:
- github_listing.go  — ListIssuesByStates + the open-PR claim scan + paging.
- github_mapping.go  — issue/label mapping, state-query helpers, cap-hit logging.
- github_refresh.go  — FetchIssueStatesBy{IDs,Refs} narrow state refresh + number cache.

Verified lossless: the top-level declaration set is byte-identical before/after
(47 decls), the moved function bodies diff to zero, and github.go's import block
is the goimports-trimmed subset the core uses. github.go 772 -> 114; every file
is now well under the 800 cap. No public API added for the budget; package
tracker is unchanged externally.

This is the first of #821's two files (its own PR per the acceptance criteria);
tracker_client.go is tracked separately.

Closes #821
